# cuda : improve text-generation and batched decoding performance