--- a/ggml-cuda.cu
+++ b/ggml-cuda.cu
@@ -92,6 +92,7 @@
 // for large computational tasks. the drawback is that this requires some extra amount of VRAM:
 // -  7B quantum model: +100-200 MB
 // - 13B quantum model: +200-400 MB
+//
 //#define GGML_CUDA_FORCE_MMQ
 
 // TODO: improve this to be correct for more hardware