ref #3479 
ref #3771 

# Description

This PR should improve significantly the text-generation, prompt processing and batched decoding speed for all models for NVIDIA cards with tensor cores (i.e. VOLTA, AMPERE, etc).

## Prompt processing

By default `llama.cpp` uses `MMQ=1` which means that the matrix-matrix multiplications for quantized models are performed with custom kernel for integer multiplications. Recently (#3412), we found out that for large batch dimension (which is the case when processing prompts), `MMQ=0` offers significant performance boost by first dequantizing `src0` to F16 and performing the GEMM using cublas. This PR essentially enables the same optimization for `MMQ=1` by not using the custom kernel for batch size > 32.

## Batched decoding

In this mode, the batch size is larger than 1, but typically small (for example not more than 32). In #3545 we found out that the currently used constants `MMQ_X`, `MMQ_Y` and `NWARPS` are not optimal for small batch sizes. Probably they have been optimized for prompt processing. However, since we now fallback to cuBLAS for prompt processing, the constants can be adjusted for small batch sizes.

## Text-generation

So far, for the KV cache related ops (`KQ` and `KQV`) we have been using custom matrix-vector kernels. For small sequence lengths (~128) and no prompt, these kernels are quite efficient. However, as the KV cache grows with the sequence length it is more efficient to use the tensor cores via cuBLAS GEMM. This PR applies this change to achieve TG improvements for all models when the context is big

In summary, we now have the following strategy for matrix multiplications:

- batch size == 1:
  - non-attention ops && quantized `src0`: use custom matrix-vector kernel
  - otherwise: use cuBLAS GEMM
- batch size <= 32: use custom matrix-matrix kernel
- batch size > 32: use cuBLAS GEMM

# Results

## RTX 3090

```bash
LLAMA_CUBLAS=1 make -j batched batched-bench && ./batched-bench ./models/codellama-7b/ggml-model-q4_0.gguf 8704 1 99 1 512 128 1,2,3,4,5,6,7,8,16,32,64
```

### master

main: n_kv_max = 8704, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.248 |  2063.97 |    1.098 |   116.60 |    1.346 |   475.54 |
|   512 |    128 |    2 |    768 |    0.234 |  2191.11 |    7.435 |    34.43 |    7.669 |   100.15 |
|   512 |    128 |    3 |    896 |    0.233 |  2198.40 |    7.515 |    51.09 |    7.748 |   115.64 |
|   512 |    128 |    4 |   1024 |    0.234 |  2189.40 |    7.553 |    67.78 |    7.787 |   131.50 |
|   512 |    128 |    5 |   1152 |    0.236 |  2168.46 |    7.635 |    83.83 |    7.871 |   146.37 |
|   512 |    128 |    6 |   1280 |    0.236 |  2167.85 |    7.700 |    99.74 |    7.936 |   161.28 |
|   512 |    128 |    7 |   1408 |    0.238 |  2152.19 |    7.768 |   115.34 |    8.006 |   175.86 |
|   512 |    128 |    8 |   1536 |    0.242 |  2115.82 |    7.832 |   130.75 |    8.074 |   190.25 |
|   512 |    128 |   16 |   2560 |    0.244 |  2095.92 |    8.391 |   244.06 |    8.636 |   296.45 |
|   512 |    128 |   32 |   4608 |    0.264 |  1937.69 |    8.852 |   462.71 |    9.116 |   505.46 |
|   512 |    128 |   64 |   8704 |    0.245 |  2085.75 |   11.646 |   703.42 |   11.891 |   731.95 |

### PR

main: n_kv_max = 8704, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.173 |  2953.07 |    1.073 |   119.32 |    1.246 |   513.57 |
|   512 |    128 |    2 |    768 |    0.165 |  3109.78 |    1.715 |   149.23 |    1.880 |   408.49 |
|   512 |    128 |    3 |    896 |    0.162 |  3155.39 |    1.768 |   217.16 |    1.931 |   464.11 |
|   512 |    128 |    4 |   1024 |    0.164 |  3123.65 |    1.803 |   283.98 |    1.967 |   520.63 |
|   512 |    128 |    5 |   1152 |    0.163 |  3138.81 |    2.403 |   266.36 |    2.566 |   448.96 |
|   512 |    128 |    6 |   1280 |    0.166 |  3077.79 |    2.433 |   315.70 |    2.599 |   492.49 |
|   512 |    128 |    7 |   1408 |    0.166 |  3082.24 |    2.515 |   356.21 |    2.681 |   525.08 |
|   512 |    128 |    8 |   1536 |    0.166 |  3080.13 |    2.551 |   401.48 |    2.717 |   565.37 |
|   512 |    128 |   16 |   2560 |    0.169 |  3028.26 |    4.723 |   433.60 |    4.892 |   523.27 |
|   512 |    128 |   32 |   4608 |    0.167 |  3059.51 |    7.973 |   513.75 |    8.140 |   566.09 |
|   512 |    128 |   64 |   8704 |    0.164 |  3126.79 |   11.462 |   714.74 |   11.625 |   748.71 |

```bash
LLAMA_CUBLAS=1 make -j batched-bench && ./batched-bench ./models/codellama-13b/ggml-model-q4_k.gguf 4096 1 99 1 512,3200 128,800 1
```

### master

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.437 |  1171.55 |    2.124 |    60.28 |    2.561 |   249.95 |
|   512 |    800 |    1 |   1312 |    0.435 |  1176.46 |   14.269 |    56.07 |   14.704 |    89.23 |
|  3200 |    128 |    1 |   3328 |    3.233 |   989.79 |    3.422 |    37.40 |    6.655 |   500.07 |
|  3200 |    800 |    1 |   4000 |    3.219 |   994.22 |   22.392 |    35.73 |   25.611 |   156.19 |

### PR

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.246 |  2084.35 |    2.021 |    63.32 |    2.267 |   282.30 |
|   512 |    800 |    1 |   1312 |    0.247 |  2074.85 |   13.436 |    59.54 |   13.683 |    95.88 |
|  3200 |    128 |    1 |   3328 |    1.977 |  1618.55 |    2.501 |    51.18 |    4.478 |   743.16 |
|  3200 |    800 |    1 |   4000 |    1.984 |  1613.30 |   16.140 |    49.57 |   18.123 |   220.71 |

```bash
make -j && ../scripts/run-all-perf.sh codellama-7b "f16 q8_0 q4_0 q4_k" "-ngl 999 -t 1 -n 128,512 -p 512"
```

### master

| model                   |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ----------------------- | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  4493.64 ± 25.65 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |     54.67 ± 0.04 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     52.65 ± 0.24 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  2038.44 ± 73.46 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |     87.10 ± 0.09 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     82.33 ± 0.25 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     | 1901.29 ± 154.12 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    133.48 ± 0.17 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    123.13 ± 0.38 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  1998.98 ± 52.16 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    110.47 ± 0.18 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    101.51 ± 0.29 |

### PR

| model                   |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ----------------------- | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  4513.66 ± 32.46 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |     54.14 ± 0.13 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     53.00 ± 0.26 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  3183.63 ± 16.25 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |     85.08 ± 0.15 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     82.86 ± 0.14 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  3250.48 ± 22.76 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    130.02 ± 0.09 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    126.12 ± 0.29 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  3661.56 ± 53.75 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    109.18 ± 0.19 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    104.73 ± 0.08 |

## RTX 4090

```bash
LLAMA_CUBLAS=1 make -j batched batched-bench && ./batched-bench ./models/codellama-7b/ggml-model-q4_0.gguf 8704 1 99 1 512 128 1,2,3,4,5,6,7,8,16,32,64
```

### master

main: n_kv_max = 8704, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.142 |  3615.61 |    0.915 |   139.85 |    1.057 |   605.55 |
|   512 |    128 |    2 |    768 |    0.107 |  4779.37 |    5.170 |    49.52 |    5.277 |   145.55 |
|   512 |    128 |    3 |    896 |    0.126 |  4050.31 |    5.201 |    73.83 |    5.328 |   168.17 |
|   512 |    128 |    4 |   1024 |    0.120 |  4268.98 |    5.210 |    98.27 |    5.330 |   192.12 |
|   512 |    128 |    5 |   1152 |    0.128 |  3988.04 |    5.240 |   122.15 |    5.368 |   214.60 |
|   512 |    128 |    6 |   1280 |    0.126 |  4062.23 |    5.242 |   146.51 |    5.368 |   238.45 |
|   512 |    128 |    7 |   1408 |    0.138 |  3721.15 |    5.336 |   167.91 |    5.474 |   257.22 |
|   512 |    128 |    8 |   1536 |    0.114 |  4493.71 |    5.302 |   193.12 |    5.416 |   283.59 |
|   512 |    128 |   16 |   2560 |    0.115 |  4435.90 |    5.558 |   368.49 |    5.673 |   451.24 |
|   512 |    128 |   32 |   4608 |    0.118 |  4325.20 |    5.853 |   699.78 |    5.972 |   771.65 |
|   512 |    128 |   64 |   8704 |    0.120 |  4251.54 |    7.116 |  1151.21 |    7.236 |  1202.81 |

### PR

main: n_kv_max = 8704, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.109 |  4697.51 |    0.943 |   135.73 |    1.052 |   608.36 |
|   512 |    128 |    2 |    768 |    0.089 |  5738.82 |    1.389 |   184.31 |    1.478 |   519.57 |
|   512 |    128 |    3 |    896 |    0.099 |  5188.49 |    1.410 |   272.26 |    1.509 |   593.72 |
|   512 |    128 |    4 |   1024 |    0.091 |  5633.43 |    1.438 |   355.94 |    1.529 |   669.57 |
|   512 |    128 |    5 |   1152 |    0.093 |  5476.76 |    1.508 |   424.27 |    1.602 |   719.12 |
|   512 |    128 |    6 |   1280 |    0.086 |  5968.90 |    1.520 |   505.42 |    1.605 |   797.35 |
|   512 |    128 |    7 |   1408 |    0.092 |  5567.34 |    1.546 |   579.54 |    1.638 |   859.57 |
|   512 |    128 |    8 |   1536 |    0.091 |  5653.09 |    1.574 |   650.69 |    1.664 |   922.91 |
|   512 |    128 |   16 |   2560 |    0.084 |  6129.17 |    2.196 |   932.58 |    2.280 |  1123.01 |
|   512 |    128 |   32 |   4608 |    0.099 |  5172.66 |    3.436 |  1192.04 |    3.535 |  1303.50 |
|   512 |    128 |   64 |   8704 |    0.097 |  5279.28 |    6.336 |  1293.00 |    6.433 |  1353.10 |


```bash
LLAMA_CUBLAS=1 make -j batched-bench && ./batched-bench ./models/codellama-13b/ggml-model-q4_k.gguf 4096 1 99 1 512,3200 128,800 1
```

### master

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.226 |  2269.96 |    1.607 |    79.66 |    1.832 |   349.29 |
|   512 |    800 |    1 |   1312 |    0.214 |  2387.60 |   10.557 |    75.78 |   10.771 |   121.80 |
|  3200 |    128 |    1 |   3328 |    1.640 |  1950.80 |    2.294 |    55.80 |    3.934 |   845.87 |
|  3200 |    800 |    1 |   4000 |    1.626 |  1968.49 |   14.876 |    53.78 |   16.501 |   242.41 |

### PR

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.152 |  3363.09 |    1.612 |    79.40 |    1.764 |   362.73 |
|   512 |    800 |    1 |   1312 |    0.145 |  3522.46 |   10.201 |    78.42 |   10.347 |   126.80 |
|  3200 |    128 |    1 |   3328 |    1.269 |  2521.20 |    1.986 |    64.45 |    3.255 |  1022.30 |
|  3200 |    800 |    1 |   4000 |    1.268 |  2523.72 |   12.692 |    63.03 |   13.960 |   286.53 |

```bash
make -j && ../scripts/run-all-perf.sh codellama-7b "f16 q8_0 q4_0 q4_k" "-ngl 999 -t 1 -n 128,512 -p 512"
```

### master

| model                   |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ----------------------- | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     | 8972.06 ± 345.72 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |     62.27 ± 0.19 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     61.54 ± 0.04 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |  5272.91 ± 99.01 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    101.73 ± 0.15 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     99.43 ± 0.05 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     | 5017.87 ± 132.87 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    158.38 ± 0.13 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    152.41 ± 0.85 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     | 4763.11 ± 163.93 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    147.54 ± 0.30 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    142.81 ± 0.15 |

### PR

| model                   |       size |     params | backend    | ngl |    threads | test       |              t/s |
| ----------------------- | ---------: | ---------: | ---------- | --: | ---------: | ---------- | ---------------: |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     | 9473.31 ± 227.98 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |     61.76 ± 0.29 |
| llama 7B F16            |  12.55 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     61.27 ± 0.01 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |   6473.99 ± 4.55 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    100.22 ± 0.14 |
| llama 7B Q8_0           |   6.67 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |     98.65 ± 0.07 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |   6693.25 ± 6.18 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    154.33 ± 0.36 |
| llama 7B Q4_0           |   3.56 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    150.87 ± 0.12 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | pp 512     |   7277.83 ± 4.73 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 128     |    144.40 ± 0.25 |
| llama 7B Q4_K - Medium  |   3.80 GiB |     6.74 B | CUDA       | 999 |          1 | tg 512     |    141.19 ± 0.07 |


## V100

```bash
LLAMA_CUBLAS=1 make -j batched-bench && ./batched-bench ./models/openllama-7b-v2/ggml-model-q4_k.gguf 4096 1 99 1 512,3200 128,800 1
```

### master

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.353 |  1451.95 |    1.388 |    92.23 |    1.740 |   367.73 |
|   512 |    800 |    1 |   1312 |    0.351 |  1457.49 |    9.431 |    84.83 |    9.782 |   134.12 |
|  3200 |    128 |    1 |   3328 |    2.648 |  1208.35 |    2.329 |    54.97 |    4.977 |   668.71 |
|  3200 |    800 |    1 |   4000 |    2.653 |  1206.38 |   15.285 |    52.34 |   17.937 |   223.00 |

### PR

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.211 |  2421.42 |    1.376 |    92.99 |    1.588 |   403.05 |
|   512 |    800 |    1 |   1312 |    0.212 |  2419.59 |    9.145 |    87.48 |    9.357 |   140.22 |
|  3200 |    128 |    1 |   3328 |    1.767 |  1811.34 |    2.026 |    63.17 |    3.793 |   877.43 |
|  3200 |    800 |    1 |   4000 |    1.765 |  1812.75 |   13.127 |    60.94 |   14.892 |   268.60 |

## A100 80GB

```bash
LLAMA_CUBLAS=1 make -j batched batched-bench && ./batched-bench ./models/codellama-7b/ggml-model-q4_0.gguf 8704 1 99 1 512 128 1,2,3,4,5,6,7,8,16,32,64
```

### master

main: n_kv_max = 8704, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.234 |  2185.63 |    1.031 |   124.18 |    1.265 |   505.93 |
…