@JohannesGaessler 2023-10-25
If this PR does what I think it does I very much do not agree with it. Half the motivation behind mmq was to reduce the VRAM usage by avoiding the allocation of a temporary buffer for the dequantized weight matrix. Even on relatively recent hardware where the speed is now comparatively worse this can be worthwhile if VRAM is tight.

Also as far as I can tell there is no check for CUDA architectures without tensor cores where mmq should be universally faster or AMD GPUs (where I don't know hich one is faster).