@ggerganov 2023-10-25
> What is the reason for adding a new function 

Likely no reason, but I wanted to eliminate the convoluted logic of `ggml_cuda_op_mul_mat` in the process - it was just easier to work this way. I've added a TODO to remove the duplicated call

> Also as far as I can tell there is no check for CUDA architectures without tensor cores where mmq should be universally faster or for AMD GPUs (where I don't know which one is faster).

I will try to add an elegant way to accommodate old video cards and fallback to MMQ (should be easy), but for sure we are not going to hold back performance on modern hardware where there is plenty of VRAM just to be able to offload a few more layers. So this change in one way or another is going to make it to `master` unless there is some bug of course.