@ggerganov 2023-10-25
What do you propose?

With this change:

We gain:
- 1.5x PP speed
- 5x Batched decoding speed needed for real-world applications and speculative decoding

We lose:
- ~100 MB VRAM for quantum 7B models
- ~200 MB VRAM for quantum 13B models