@JohannesGaessler 2023-10-25
>Ultimately, the best solution would be to store activations as F16 and implementing our own MMQ kernels using tensor cores. This would remove the need for these buffers entirely.

I agree that long-term this would be the best solution. However, using tensor cores on their own does **not** seem to be sufficient. Several weeks ago I did a prototype mmq implementation using tensor cores and there was no speedup because the tensor cores were dramatically underutilized. In fact, even without the use of tensor cores the ALU using the current mmq utilization is underutilized. The problem is most likely related to mmq not utilizing asynchronous data loading (available on Ampere or newer). For an mmq implementation that outperforms FP16 cuBLAS for quantized models I think the following changes would be necessary:

1. Implement an mmq kernel that loads the data asynchronously.
2. Modify the mmq kernels to make the tiles thinner in direction of `src0` and `src1` dimension 0 through another template parameter. The tradeoff would be potentially worse data coalescing but reduced shared memory usage. With thinner tiles you should be able to increase the `MMQ_X` and `MMQ_Y` parameters which in turn should increase the arithmetic intensity. Also there would need to be special consideration for k-quants since currently the tiles are always at least as big as a single k-quant block.
3. Utilize tensor cores since now I/O is less of a bottleneck.

Currently it looks like I will be busy until the end of December. Afterwards I wanted to start working on this unless someone else does something similar in the meantime.