@ggerganov 2023-10-25
> Ultimately, the best solution would be to store activations as F16 and implementing our own MMQ kernels using tensor cores.

I agree that this is the right direction and we will do it eventually. It's just that this change is so easy (if you remove the duplicated code it's just a few lines) that I don't see any reason to not get the benefits from it.

> When offloading the output tensor

I just realized the output tensor currently does not go through the new cublas GEMM branch because I use `all_on_device` and the dst is not on device. Will see how it affects the performance and depending on the result, we can leave it as it is.

@JohannesGaessler When I wrote earlier that I will try to accommodate older cards and fallback to MMQ, I had something similar in mind to what you propose.