@ggerganov 2023-10-26
Using `cublasGemmBatchedEx` for single-batch (i.e. text-generation mode), slightly degrades the TG speed for short sequence length, but significantly improves the performance for long sequence length (i.e. when the context is big). So I've enabled this as well when tensor cores are available.