@oobabooga 2023-10-26
I built a custom version of llama-cpp-python using this PR branch and re-did the tests in [this post](https://oobabooga.github.io/blog/posts/gptq-awq-exl2-llamacpp/) for `llama-2-13b-Q4_K_M.gguf`. These are the results using my RTX 3090:

| llama.cpp                                   | Prompt Processing Time (3200 tokens, in seconds) | Evaluation Time (800 tokens, in seconds) | Tokens/Second | VRAM (MB) |
|----------------------------------------|---------------------------------------|------------------------------|--------------|------|
| llama-cpp-python 0.2.11                | 3.73                                | 25.95                        | 30.83        | 8985 |
| ggerganov PR branch | 2.19                                | 16.44                        | 48.66        | 8983 |

This is a +58% increase in the tokens/second during evaluation, and a +70% increase in the tokens/second during processing, at no VRAM cost (in my test).