@JohannesGaessler 2023-10-26
Sorry, but are you sure you tested that correctly? This PR only affects batch sizes >1 so there should be no difference whatsoever when generating one token at a time.

Also, if you are measuring VRAM by what is printed to console that is **not** reflective of the actual VRAM usage. The print only includes the VRAM allocated initially and does not account for the additional VRAM potentially needed later when it may be possible to dequantize a weight matrix to either FP16 or FP32 (i.e. when mmq is not used). Instead something like nvidia-smi should be used to monitor the VRAM usage.