@Dampfinchen 2023-10-26
I will give my input here. Mind you, this is before I got the time to test this PR:

With MMQ I am able to run 7B Q4K_S models with GQA like Mistral fully offloaded on my RTX 2060 with 6 GB VRAM and at 4K context. And I still have some VRAM left to spare for other applications.

With CUDA on master, it will slow down fast as it overflows into system ram. 

So for me, the VRAM savings MMQ in master provides are essential. I hope MMQ will continue to stay VRAM efficient. 

I do hope tensor cores can be used for it eventually and be super VRAM efficient at the same time. CUDA is noticeably faster (when the VRAM is not spilling over)